tests.data: mock processors and ocrd-tool.json
