+++ tests/processor/test_processor.py
@@ -3,46 +3,10 @@ import json
 from tempfile import TemporaryDirectory
 from os.path import join
 from tests.base import TestCase, assets, main # pylint: disable=import-error, no-name-in-module
+from tests.data import DummyProcessor, DummyProcessorWithRequiredParameters, IncompleteProcessor, DUMMY_TOOL
 
 from ocrd.resolver import Resolver
 from ocrd.processor.base import Processor, run_processor, run_cli
-
-DUMMY_TOOL = {
-    'executable': 'ocrd-test',
-    'steps': ['recognition/post-correction'],
-    'parameters': {
-        'baz': {
-            'type': 'string',
-            'default': 'bla'
-        }
-    }
-}
-
-class DummyProcessor(Processor):
-
-    def __init__(self, *args, **kwargs):
-        kwargs['ocrd_tool'] = DUMMY_TOOL
-        kwargs['version'] = '0.0.1'
-        super(DummyProcessor, self).__init__(*args, **kwargs)
-
-    def process(self):
-        print(json.dumps(self.parameter))
-
-class DummyProcessorWithRequiredParameters(Processor):
-    def process(self): pass
-    def __init__(self, *args, **kwargs):
-        kwargs['version'] = '0.0.1'
-        kwargs['ocrd_tool'] = {
-            'executable': 'ocrd-test',
-            'steps': ['recognition/post-correction'],
-            'parameters': {
-                'i-am-required': {'required': True}
-            }
-        }
-        super(DummyProcessorWithRequiredParameters, self).__init__(*args, **kwargs)
-
-class IncompleteProcessor(Processor):
-    pass
 
 class TestProcessor(TestCase):
